split print.c
